Renaming of getter classes
