+++ lib/commit_lister/git_cli/git_cli_wrapper.rb
